Split DesignControl class files
* Split DesignControl class into partial class files to separate the
major functionality of each of its components.
* Refactor WinAPIMethods to WinApiMethods.
* Refactor some methods in DesignForm to be more code-independent.
* Remove DrawHelper class.
* Add generic EventArgs class.
* Add null check in RtlAwareMessageBox class when checking for
RightToLeft.
* Split TileMapControl class into partial classes.
* Style changes to TileMapControl1D class code.
* Update Size adjust methods in TileMapForm class to be event-based.
* Change UnsafeNativeMethods class access from public to internal.
* Add IEquatable interface to WinApiRectangle.
