+++ src/Controls/Controls.csproj
@@ -35,6 +35,7 @@
   <ItemGroup>
     <Reference Include="System" />
     <Reference Include="System.Core" />
+    <Reference Include="System.Design" />
     <Reference Include="System.Drawing" />
     <Reference Include="System.Windows.Forms" />
     <Reference Include="System.Xml.Linq" />
@@ -53,6 +54,15 @@
       <SubType>UserControl</SubType>
     </Compile>
     <Compile Include="DashedPenPair.cs" />
+    <Compile Include="DesignControl.Drawing.cs">
+      <SubType>UserControl</SubType>
+    </Compile>
+    <Compile Include="DesignControl.Keys.cs">
+      <SubType>UserControl</SubType>
+    </Compile>
+    <Compile Include="DesignControl.Mouse.cs">
+      <SubType>UserControl</SubType>
+    </Compile>
     <Compile Include="DialogForm.cs">
       <SubType>Form</SubType>
     </Compile>
@@ -63,7 +73,7 @@
       <SubType>Form</SubType>
     </Compile>
     <Compile Include="DialogProxy.cs" />
-    <Compile Include="DrawHelper.cs" />
+    <Compile Include="EventArgs.cs" />
     <Compile Include="IDialogForm.cs" />
     <Compile Include="IIntegerComponent.cs" />
     <Compile Include="IntegerTextBox.cs">
@@ -80,6 +90,12 @@
     </Compile>
     <Compile Include="RtlAwareMessageBox.cs" />
     <Compile Include="TileMapControl.cs">
+      <SubType>UserControl</SubType>
+    </Compile>
+    <Compile Include="TileMapControl.Scroll.cs">
+      <SubType>UserControl</SubType>
+    </Compile>
+    <Compile Include="TileMapControl.TileMap.cs">
       <SubType>UserControl</SubType>
     </Compile>
     <Compile Include="TileMapControl1D.cs">
